requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-cors-headers==4.7.0
django-stubs-ext==5.1.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fire==0.7.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
kombu==5.5.2
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
termcolor==3.0.1
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250328
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
